Add an fstab nofail mount option
nofail mounts won't prevent the system from booting if not mountable.
systemd will stil try to mount them in an asynchronous way, and add an
error to the log if it fails.
Most mount points, apart from / and /home should be mounted with nofail
to prevent entering to recovery mode with a bootable system.
partitionmanager doesn't remove fstab entries after removing a
partition, which currently leads to recovery mode on the next boot.
BUG: 472431